dev-requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile dev-requirements.in -o dev-requirements.txt
iniconfig==2.0.0
    # via pytest
nodeenv==1.8.0
    # via pyright
numpy==2.3.5
    # via pandas-stubs
packaging==24.0
    # via pytest
pandas-stubs==2.3.2.250926
    # via -r dev-requirements.in
pluggy==1.5.0
    # via pytest
psutil==7.1.3
    # via -r dev-requirements.in
pyright==1.1.360
    # via -r dev-requirements.in
pytest==8.1.1
    # via -r dev-requirements.in
ruff==0.4.2
    # via -r dev-requirements.in
setuptools==69.5.1
    # via
    #   -r dev-requirements.in
    #   nodeenv
types-pytz==2025.2.0.20251108
    # via pandas-stubs
